Performance Analysis of MongoDB and HBase
Installed and implemented instances of MongoDB and HBase. Implemented a test strategy in order to perform a comparative analysis of the capabilities of each system in terms of the performance. Recorded a set of appropriate metrics and performed a quantitative analysis for comparison purposes between the two chosen systems. Comparison done on the basis of security level.